# Optimize the DCT GPU kernel.