#### Why we need this PR?
- Refactoring to improve the performance of the GPU DCT kernel. In the current state it's extremely slow for the transform done on the inner axis, so I handle this case separately. Also the existing CUDA kernel was slightly optimized.

#### What happened in this PR?
 - What solution was applied:
     The case with the transform done over the inner axis is handled with a separate CUDA kernel. The existing kernel was optimized by employing shared memory.
 - Affected modules and functionalities:
    GPU DCT kernel.
 - Key points relevant for the review:
     A new CUDA kernel and the changes in the old one.
 - Validation and testing:
     Existing tests still apply. I added a performance test.
 - Documentation (including examples):
     N/A


**JIRA TASK**: *DALI-1690*