@JanuszL 2020-11-17
What is the speed now, and what used to be before that optimization?