@banasraf 2020-11-17
@JanuszL 
> What is the speed now, and what used to be before that optimization?
For the planar layout it's ~550 GFLOPS -> ~630 GFLOPS and for interleaved it's ~30 GFLOPS -> ~315 GFLOPS